Extract call-argument initializer fixtures
